+++ Cargo.toml
@@ -3,6 +3,7 @@ resolver = "3"
 members = [
     "crates/attestation",
     "crates/attestation-cli",
+    "crates/attestation-types",
     "crates/backup-cli",
     "crates/chain-gateway",
     "crates/chain-gateway-test-contract",
@@ -48,6 +49,7 @@ repository = "https://github.com/near/mpc"
 [workspace.dependencies]
 #workspace members
 attestation = { path = "crates/attestation" }
+attestation-types = { path = "crates/attestation-types" }
 chain-gateway = { path = "crates/chain-gateway" }
 chain-gateway-test-contract = { path = "crates/chain-gateway-test-contract" }
 contract-history = { path = "crates/contract-history" }
