+++ crates/attestation-types/assets/event_log.json
@@ -1,0 +1,7 @@
+{
+    "imr": 0,
+    "event_type": 2147483659,
+    "digest": "8ae1e425351df7992c444586eff99d35af3b779aa2b0e981cb4b73bc5b279f2ade19b6a62a203fc3c3bbdaae80af596d",
+    "event": "",
+    "event_payload": "095464785461626c65000100000000000000af96bb93f2b9b84e9462e0ba745642360090800000000000"
+}
